requirements.txt
asgiref==3.7.2
Django==5.0.1
django-modeltranslation==0.18.11
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
